+++ survey/reports.py
@@ -1,0 +1,19 @@
+
+from django.db.models import Count
+
+from .models import CoachSurvey, CoachSurveySubmission, CoachSurveySubmissionDraft
+
+
+def survey_aggregates():
+    """
+    Required fields:
+        - Name of survey
+        - Date published
+        - Total users completed
+        - Total users who started but did not complete the survey
+        - Total users who choose not to consent
+        - Total user who did not respond
+    """
+    # TODO: Survey Aggregate Report
+
+    pass