+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+0.1.1
+-----
+
+- Implemented Survey drafts
+- Aggregating users who have selected a Goal Prototype 
+
 0.0.15
 ------